Git: Commit : Error Handling

+++ src/main/java/com/blog/blogSite/service/AuthorServiceImpl.java
@@ -1,12 +1,16 @@
 package com.blog.blogSite.service;
 
 import com.blog.blogSite.entity.Author;
+import com.blog.blogSite.entity.Comment;
+import com.blog.blogSite.exception.ResourceAlreadyExists;
+import com.blog.blogSite.exception.ResourceNotFoundException;
 import com.blog.blogSite.repository.AuthorRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 /**
  * Author service implementation
@@ -21,6 +25,11 @@ public class AuthorServiceImpl implements AuthorService
 
     @Override
     public Author saveAuthor(Author author) {
+        Optional<Author> byId = authorRepository.findById(author.getId());
+        if (byId.isPresent())
+        {
+            throw new ResourceAlreadyExists();
+        }
         return authorRepository.save(author);
     }
 
@@ -31,11 +40,21 @@ public class AuthorServiceImpl implements AuthorService
 
     @Override
     public Author fetchAuthorsById(Long authorId) {
+        Optional<Author> byId = authorRepository.findById(authorId);
+        if (byId.isEmpty())
+        {
+            throw new ResourceNotFoundException();
+        }
         return  authorRepository.findById(authorId).get();
     }
 
     @Override
     public Author updateAuthor(Author author, Long authorId) {
+        Optional<Author> byId = authorRepository.findById(authorId);
+        if (byId.isEmpty())
+        {
+            throw new ResourceNotFoundException();
+        }
         Author authorDB = authorRepository.findById(authorId).get();
 
         if (Objects.nonNull(author.getName()) && !"".equalsIgnoreCase(author.getName())) {
@@ -59,6 +78,11 @@ public class AuthorServiceImpl implements AuthorService
 
     @Override
     public void deleteAuthorById(Long authorId) {
+        Optional<Author> byId = authorRepository.findById(authorId);
+        if (byId.isEmpty())
+        {
+            throw new ResourceNotFoundException();
+        }
         authorRepository.deleteById(authorId);
     }
