+++ src/main/java/com/blog/blogSite/service/PostServiceImpl.java
@@ -1,12 +1,16 @@
 package com.blog.blogSite.service;
 
+import com.blog.blogSite.entity.Comment;
 import com.blog.blogSite.entity.Post;
+import com.blog.blogSite.exception.ResourceAlreadyExists;
+import com.blog.blogSite.exception.ResourceNotFoundException;
 import com.blog.blogSite.repository.PostRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 /**
  * Post service implementation
@@ -20,6 +24,11 @@ public class PostServiceImpl implements PostService
 
     @Override
     public Post savePost(Post post) {
+        Optional<Post> byId = postRepository.findById(post.getId());
+        if (byId.isPresent())
+        {
+            throw new ResourceAlreadyExists();
+        }
         return postRepository.save(post);
     }
 
@@ -30,11 +39,21 @@ public class PostServiceImpl implements PostService
 
     @Override
     public Post fetchPostById(Long postId) {
+        Optional<Post> byId = postRepository.findById(postId);
+        if (byId.isEmpty())
+        {
+            throw new ResourceNotFoundException();
+        }
         return postRepository.findById(postId).get();
     }
 
     @Override
     public Post updatePost(Post post, Long postId) {
+        Optional<Post> byId = postRepository.findById(postId);
+        if (byId.isEmpty())
+        {
+            throw new ResourceNotFoundException();
+        }
         Post postDB = postRepository.findById(postId).get();
 
         if (Objects.nonNull(post.getTitle()) && !"".equalsIgnoreCase(post.getTitle())) {
@@ -62,6 +81,11 @@ public class PostServiceImpl implements PostService
 
     @Override
     public void deletePostById(Long postId) {
+        Optional<Post> byId = postRepository.findById(postId);
+        if (byId.isEmpty())
+        {
+            throw new ResourceNotFoundException();
+        }
         postRepository.deleteById(postId);
     }
 }
